Building a tree now works.
Any arbitrary peer can be the root of the tree. The tree building avoids cycles.
This means that, in reality, the tree representation may not be the true form
of the connections. In the future, broadcasters will likely form a fully connected
mesh. To downstream peers, they will appear like a linked list, but internally they
are not. These special cases are yet to be handled. Also, peer selection does not
use the tree.